Take into account which button was pressed
Why: ControlUnit needs to know which process needs to be executed.

This change addresses the need by: Checking what was sent from the
SimulatorWindow. If SimulatorWindow passed "Single Step" then only one
Instruction needs to be executed. Else if it was "Execute" then the
ControlUnit needs to start the fetch-execute cycle.